backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
opencv-python==4.9.0.80
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
python-multipart==0.0.7
websockets==12.0
twilio==8.12.0
cryptography==42.0.1
pillow==10.2.0
numpy==1.26.3
ffmpeg-python==0.2.0
reportlab==4.1.0
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-jose==3.3.0
httpx==0.26.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
